scp-github/python/scp-router-personal/requirements.txt
--only-binary :platform_system!="Windows"  # For scikit-learn

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# HTTP Client
httpx==0.25.1

# Database & Cache
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Lightweight text processing
scikit-learn==1.3.0  # For TF-IDF (lightweight, no torch)
numpy>=1.26.0 

# LLM API Client Claude
anthropic==0.86.0